Editing add to Cart

+++ database/migrations/2020_09_02_175111_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('price');
+            $table->unsignedBigInteger('price');
             $table->boolean('available');
             $table->string('image_path')->nullable();
             $table->unsignedBigInteger('quantity')->default(1);
